Refactor AI Agent tools to modular architecture
- Created `AgentTool` interface and categorized tools (DATA, UTILITY, ACTION)
- Extracted schemas and logic from `AgentTools.kt` and `ToolRegistry.kt` into individual classes (e.g., `AddTaskTool`, `GamificationTool`)
- Cleaned up giant `when` blocks in execution layer
- Updated `AISettingsActivity` to group tools visually by category
- Verified cross-package compatibility and test integrity
